+++ Server/MoneyManager/Model/Common/UpdateAccountModel.cs
@@ -8,6 +8,6 @@ namespace MoneyManager.Model.Common
     {
         public Guid AccountId { get; set; }
 
-        public double Delta { get; set; }
+        public decimal Delta { get; set; }
     }
 }
